Reorganize and document test file groupings
