+++ tests/test_individual_citation_fns.py
@@ -1,5 +1,4 @@
 import citationAnalysis.citationAnalysis as ca
-from UI.fileManipulation import find_citations
 
 # The \ indicates code continues in the next line
 # When testing individual detection functions, they SHOULD catch extra characters
@@ -173,13 +172,3 @@ def test_surnames_apostrophe_s_recognised():
     matches.cleanup()
     assert matches.citations == ["O'Samuel 1999", "O'Sullivan 2000"]
  
-def test_find_citations_wrapper():
-    text = "One and two (1212) often collaborate. This is well documented (Aesop and Berry, 1999). Aasimov (1999) had a revelation (Bassimov, 2000). Long Surname (2020) found his match."
-    
-    # Specifying sought-after matches. Berry, 1999 is a side-effect not considered harmful.
-    expected_narrow = ["Aasimov 1999", "Bassimov 2000", "Aesop & Berry 1999", "One & two 1212"]
-    expected_wider = ["Long Surname 2020"]
-    
-    found_citations = find_citations(text)
-    assert all([citation in found_citations[0].citations for citation in expected_narrow])
-    assert all([citation in found_citations[1].citations for citation in expected_wider])
